ifxed

+++ src/app/api/posts/[id]/route.ts
@@ -1,11 +1,14 @@
 import { NextRequest, NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
 
-export async function DELETE(
-  req: NextRequest,
-  context: { params: { id: string } }
-) {
-  const postId = Number(context.params.id);
+export async function DELETE( req: NextRequest ) {
+  const url = new URL(req.url);
+  const id = url.pathname.split("/").pop();
+  const postId = Number(id);
+
+  if (isNaN(postId)) {
+    return NextResponse.json({ message: "Invalid ID" }, { status: 400 });
+  }
 
   try {
     const deletedPost = await prisma.post.delete({
@@ -13,7 +16,8 @@ export async function DELETE(
     });
 
     return NextResponse.json(deletedPost);
-  } catch (error) {
+  } catch (err) {
+    console.error(err);
     return NextResponse.json(
       { message: "Post not found or already deleted" },
       { status: 404 }
